+++ CHANGELOG.rst
@@ -4,6 +4,8 @@ Changelog
 0.3.3 - unreleased
 ------------------
 
+* Test ``analog.renderers`` implementation.
+
 * Fix bug in default plaintext renderer.
 
 0.3.2 - 2014-03-02